Test mocks (#12)
* test mocks unified
