+++ .gitignore
@@ -1,5 +1,7 @@
 .DS_store
 crash.log
+.idea
+.vscode
 cloudquery
 cloudquery.db
 cloudquery.zip
@@ -8,4 +10,13 @@ dist/
 vendor/
 bin/
 .release-env
-cq-provider-k8s
+cq-provider-*
+*.exe
+*.log
+.cq_reattach
+# Local .terraform directories
+**/.terraform/*
+
+# .tfstate files
+*.tfstate
+*.tfstate.*
