+++ resources/core_nodes_test.go
@@ -5,10 +5,8 @@ import (
 
 	"github.com/cloudquery/cq-provider-k8s/client"
 	"github.com/cloudquery/cq-provider-k8s/client/mocks"
-	"github.com/cloudquery/faker/v3"
 	"github.com/golang/mock/gomock"
 	corev1 "k8s.io/api/core/v1"
-	apiresource "k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -20,51 +18,6 @@ func createCoreNodes(t *testing.T, ctrl *gomock.Controller) client.Services {
 	return client.Services{
 		Nodes: nodes,
 	}
-}
-
-func fakeNode(t *testing.T) corev1.Node {
-	// faker chokes on Node.Status.{Capacity,Allocatable} so doing it by hand
-	var node corev1.Node
-	ptrs := []interface{}{
-		&node.TypeMeta,
-		&node.ObjectMeta,
-		&node.Spec,
-		&node.Status.Phase,
-		&node.Status.Conditions,
-		&node.Status.Addresses,
-		&node.Status.DaemonEndpoints,
-		&node.Status.NodeInfo,
-		&node.Status.Images,
-		&node.Status.VolumesInUse,
-		&node.Status.VolumesAttached,
-		&node.Status.Config,
-	}
-	for i, ptr := range ptrs {
-		if err := faker.FakeData(ptr); err != nil {
-			t.Fatalf("%v %v", i, ptr)
-		}
-	}
-	rl := make(corev1.ResourceList)
-	rl["name"] = *apiresource.NewQuantity(1024*1024, apiresource.BinarySI)
-	node.Status.Capacity = rl
-	node.Status.Allocatable = rl
-	node.Spec.PodCIDR = "192.168.1.0/24"
-	node.Spec.PodCIDRs = []string{"192.168.1.0/24"}
-	node.Status.Addresses = []corev1.NodeAddress{
-		{
-			Type:    corev1.NodeHostName,
-			Address: "testname",
-		},
-		{
-			Type:    corev1.NodeInternalIP,
-			Address: "fd00::1",
-		},
-		{
-			Type:    corev1.NodeExternalIP,
-			Address: "192.168.2.1",
-		},
-	}
-	return node
 }
 
 func TestCoreNodes(t *testing.T) {
